Battery: add getVoltage()
For now, it just averages the voltage of all the modules * 2.

Note that we'd like to calculate the sum of the series pairs and then
average those, but it turns out, that's the same thing!

((a + b) + (c + d) + (e + f)) / 3 == a+b+c+d+e+f / 3
So we don't actually neeed to know about the series pairs.