State of Charge: expose directly from battery
Instead of trying to expose Amp-Hours-Remaining backward from
voltage SOC, just expose SOC and let the protocol level try to
communicate that to the inverter.

Also, stop using the expanded-capacity protocol. 6000XP doesn't seem to
support it and I couldn't get it to read the values, it always just
reported 100% no matter in what ways I tweaked the structure.