BMS: Disable charging if batter info is stale
If we lose connection to the battery for some amount of time, reality
has likely drifted from our last measurements and we should stop
charging.

An alternative would be to stop talking to the inverter and let it stop
charging, but I like the idea of the BMS making that call instead of
playing dead.