Charging: Smooth out the bumps
We can easily have oscillations in the charge current as a high current
results in a high voltage, which then results in lowering the charge
current. This can result in wild ups and downs from 100A to 0 over the
course of 10s or so.

Let's smooth it out using a simple formula: new = old * 0.9 + current * 0.1
for charge current and cell voltage range.

Note: That this formula gets applied every time the inverter asks about
the charge info (about once per 5-10s).