Charging: Stop at 100%
Previously, we'd only stop charging when hitting maxCellVolts. We'd
slowly ramp down as we approached the cell volt limit.

Let's control charging with the SOC too. IF we reach 100$, let's stop
charging for 20 min (enableCharging: false) but not affect the current.

This way, we can cut off charge current all at once but without
flipflopping around the cutoff pount.