Charging: have charge Disabled reset smoothed current to 0
Previously the smoothed "chargeCurrentLimit" wasn't set to 0
when charging was disabled. This meant that enabling / disabling
charging would happen in a big step (0 to high current in one step).

This caused oscillations of: one for 30s, off for 10min.

Now, when disabled, we set it 0 to, so when we start again, the
smoothing will slowly bring it up to the calculated level.