Pylon: Trim the incoming cell volts and temps to 15/5
The protocol isn't expecting more than this number of cells nor temps,
so let's just truncate them for now. The spec allows for a flexible
number of cells/temps, but the inverter seems to dislike getting as many
as we were giving it: 84 cell volts and 28 temps.

I'll be excited to switch protocols.