Charging Config: bring it back to safety
Let's narrow the range quite a bit to keep it in the middle.

The inverter doesn't give us control over when to stop charging, the BMS
has to determine that. But it does support setting the discharging SOC
limit. I've configured it to stop discharging at 20%.

Given I've narrowed the voltage range to 40-48, 20% should be 41.6v
which is 3.466v / cell. on the low end. The top end should be 48/12 =
4.00v per cell.