tesla modules: Stop reading faults
We don't actually use the fault system, we're constantly checking on the
battery, so let's stop asking.

This is mostly a state-machine that the BMS boards will trigger if
voltages fall outside the allowed range. Since we're checking and asking
every 10s, we don't need the system to check or flag things in-betweem.