Battery Safety: extract a tested module
I've decided that battery safety is universal and the same actions
should be take regardless of the current strategy (ramp or disable the
current as the battery nears its limits).

So, I've extracted it to a module and added some tests.

The communicated charge (or discharge) current is the MIN of both the
charging strategy and the saftey module.