Config: update volt range and charge amps
Amps: We are using a 12000XP now, which can handle double the amps.
It only claims to be able to provide 12000W (250A), but let's bump that
up a little so that at least our BMS isn't what's limits the charging.

Also, change the charging volts and the "voltsFull" value to match what
is currently deployed.