Canbus, send a full 8 byte packet for alarm data.
Some implementations had "PN" in the alarm packet, with no explantion
why. In my experimentation of trying to get comms to work I changed
this. I'm not sure if this is still needed, but since it works without
"PN" and I didn't know what that was for anyway, let's drop it.