Safety Smoothing: incorporate timing into smoothing
Previously, we were doing smoothing on each call. Now we're calling
this for RS485 charging (1 per 10s) and canbus (1 per second) so the
smoothing is way off.

Now we interpolate the smoothing exponentially based on the time.

Correspondingly, we change the smoothing to something equivalent to the
previous 0.9 in 10s (0.99 ^ 10 == 0.904)