Battery: only set the mode every so often
We expect these modes to be sticky and thus they only need to be set
when the tesla BMS boots (we think). That's only once ... possibly ever.

However, to keep it a little safer and help heal if their BMS resets or
something, let's still send these commands every 5 min.

Why: When the batttery is at high current, communication gets spotty with
the battery. Packets are corrupted mostly and thus if we send fewwr
packets, we'll need fewer successful packets to get a complete update.