Config: bump charge volts and narrow the buffer
Want to make sure the max amps for charging is what limits the current,
not the max charging voltage.

Also, drop the buffer a little so we get close to our max a little
faster. Lower the max cell volt a bit so we roll off sooner and little
faster.