Synchronization: tune for full capacity
On a synchronization day, we need to keep charging the battery till we
hit the max cell volts. So let's pull these values up a bit so we can
get a little higher.

We have one cell bank that doesn't like going very high, so it will
likely be ~50mv higher than the others.

This means that the total voltage when the highest bank hits 4.09 is:

4.09 * 11 * 4.04 = 48.53v

So, I also need to adjust our sync voltage on the shunt to match this.

In the future, we may get rid of all three of these voltages and rely
entirely on the max cell volts ceiling for: SOC, charging voltage, and
full volts.