Tesla Modules: only care about the ones in modulesInSeries
Even if it discovers extra modules, ignore them. This allows
me to disconnect one of the batteries electrically, while leaving the
chained communication cable in place.

Also, have the isFull check in getChargeDischargeInfo use our modified
SOC calcuation that tops out at 99% on sync days. Otherwise, our
charging would stop when the shunt says we are at 100%. Since the whole
point of the sync days is to reset where 100% is, we need to potentially
go *past* the possibly-drifted 100%.

Also, include the ID of modules in the history API so the entries in
`modulesInSeries` can be identified. Now that it the 'modules'
collection might be a sparse array, mapping over them collapses the
sparseness and thus we need to include the original id.