Discharging limits: add a recovery SOC %
Hitting a low-volt limit for an individual cell will cut off
discharging. Cutting off discharging eases up on the battery who's
voltage then recovers, which turns discharging back on. The result is
rapid cycling of discharging / not discharging.

Now, if we hit the lower limit, we won't re-enable discharging until SOC
recovers by a configured amount.

For now, let's set it low (1%) because we usually run out of juice only
when there's no solar coming in anyway and we want recovery to be quick
when sun rolls out.