USB Monitoring: Add script to reset / restart
USB Devices sometimes misbehave and stop communicating / working. There
are only two things you can do:

* usbreset $deviceid
   * sometimes this works (if the device is still listed)
* reboot
   * If usbreset doesn't work, or the device isn't there
     rebooting may be the only hope

So, here we make a node script (meant to be run by root) that hits the
BMS api to get the info that it needs.

This will eventually be triggered by ua cron job.