Charging / discharging: prevent rapid cycling
Charging: learn resumeChargeAtPct to only resume charging when dipping
below a certain %.

Discharging: learn resumeDischargeAtPct to only resume discharging
when charging above a certain %.

This is to prevent rapid cycling at either end of the charge cycle.

Note: we may want to reintroduce time-based sticky bools as well to
prevent rapid cycling when using voltage as a charge level instead of
the shunt.