Charging: re-add time-based charge disabling
We have SOC-based charge disabling (only re-enable when X% above the
bottom) to prevent rapid cycling. But, for the times when the shunt
is unavailable (comms sometimes go down), let's also include the
time-based charge disabling to prevent rapid cycling.

We now make charge/discharge disabling sticky for a configurable amount
of time.