Storage! Add some persistant storage
We'll definitely want to store some state that persists across reboots
of the app and machine.

First steps: Add it to latterby where it will be used.

For now, we'll just store the time since the last full charge
and do our synchronization charges based on the time since the last full
charge.

In the winter, we won't necessarily get a full charge in one day, so
this needs to stay in full-charge-mode till some system says we've fully
charged.

In particular, we will only say it's fully 100% charged if the shunt
says 100% *and* it's above our configured sync voltage. Since the shunt
is configured to call the battery 100% charged when we hit the sync
voltage, these should likely go off around the same time.