BMS: De-dupe getChargeInfo logic
We had two copies: one for RS485 and one for canbus. We don't need two
copies.

Let's replace the canbus code with the rs485 code and call the function.
There were a few recent updates to rs485 that weren't made to the canbus
code, so those updates are visible here.