Storage: validate we can read / write to the json file
Previously, if we the file didn't exist and we couldn't write to the
location, we'd load the defaults, but then fail later the first time we
wrote to the file.

Now we validate we can read/write on load. If the file is missing, we
validate we can read/write the dir.