Config: Drop the discharge by time disabling
This was initiaited to shift loads to night-time.

However, this meant that we relied only on the grid during this time.
During stormy times we want to be able to fall back to batteries and not
have discharge disabled.

Also, the peak / offpeak prices are much closer than they used to be.
Load shifting is now worth very little.